SA-1773 - Remove references to allow_report_filters_v2 feature flag
