+++ src/pages/reportBuilder/components/GroupByOption.js
@@ -1,17 +1,14 @@
 import React, { useState } from 'react';
-import { connect } from 'react-redux';
 import styles from './ReportTable.module.scss';
 import useUniqueId from 'src/hooks/useUniqueId';
 import { Box, Grid, Checkbox, Select } from 'src/components/matchbox';
 import { GROUP_CONFIG } from '../constants/tableConfig';
 import { useReportBuilderContext } from '../context/ReportBuilderContext';
-import { selectCondition } from 'src/selectors/accessConditionState';
-import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
 
 import { dehydrateFilters } from '../helpers';
 
-export const GroupByOption = props => {
-  const { groupBy, hasSubaccounts, tableLoading, _getTableData, isComparatorsEnabled } = props;
+export default function GroupByOption(props) {
+  const { groupBy, hasSubaccounts, tableLoading, _getTableData } = props;
   const { state: reportOptions } = useReportBuilderContext();
 
   const selectId = useUniqueId('break-down-by');
@@ -20,15 +17,12 @@ export const GroupByOption = props => {
 
   const handleGroupChange = e => {
     if (e.target.value !== 'placeholder') {
-      if (isComparatorsEnabled) {
-        const { filters, ...options } = reportOptions;
-        _getTableData({
-          groupBy: e.target.value,
-          reportOptions: { ...options, filters: dehydrateFilters(filters) },
-        });
-      } else {
-        _getTableData({ groupBy: e.target.value, reportOptions });
-      }
+      const { filters, ...options } = reportOptions;
+
+      _getTableData({
+        groupBy: e.target.value,
+        reportOptions: { ...options, filters: dehydrateFilters(filters) },
+      });
     }
   };
 
@@ -38,14 +32,10 @@ export const GroupByOption = props => {
     const groupBy = newTopDomainsOnly ? 'watched-domain' : 'domain';
     const { filters, ...options } = reportOptions;
 
-    if (isComparatorsEnabled) {
-      _getTableData({
-        groupBy,
-        reportOptions: { ...options, filters: dehydrateFilters(filters) },
-      });
-    } else {
-      _getTableData({ groupBy, reportOptions });
-    }
+    _getTableData({
+      groupBy,
+      reportOptions: { ...options, filters: dehydrateFilters(filters) },
+    });
   };
 
   const getSelectOptions = () => {
@@ -103,8 +93,4 @@ export const GroupByOption = props => {
       </Grid.Column>
     </Grid>
   );
-};
-
-export default connect(state => ({
-  isComparatorsEnabled: selectCondition(isAccountUiOptionSet('allow_report_filters_v2'))(state),
-}))(GroupByOption);
+}
