+++ src/pages/reportBuilder/components/ReportOptions.js
@@ -7,13 +7,12 @@ import { useReportBuilderContext } from '../context/ReportBuilderContext';
 import { selectFeatureFlaggedMetrics } from 'src/selectors/metrics';
 import { parseSearchNew as parseSearch } from 'src/helpers/reports';
 import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
-import { AddFiltersSection, CompareByForm, FiltersForm, Legend, MetricsDrawer } from './index';
+import { ActiveFilters, CompareByForm, FiltersForm, Legend, MetricsDrawer } from './index';
 import SavedReportsSection from './SavedReportsSection';
 import DateTimeSection from './DateTimeSection';
 import useRouter from 'src/hooks/useRouter';
 import { selectCondition } from 'src/selectors/accessConditionState';
 import { dehydrateFilters } from '../helpers';
-import { ActiveFilters, ActiveFiltersV2 } from './ActiveFilters';
 
 const drawerTabs = [{ content: 'Metrics' }, { content: 'Filters' }];
 export function ReportOptions(props) {
@@ -49,12 +48,8 @@ export function ReportOptions(props) {
     if (reportOptions.isReady) {
       const { filters: selectedFilters, ...update } = selectSummaryChartSearchOptions;
       const { filters } = reportOptions;
-      if (isComparatorsEnabled) {
-        update.query_filters = encodeURI(JSON.stringify(dehydrateFilters(filters)));
-      } else {
-        update.filters = selectedFilters;
-      }
 
+      update.query_filters = encodeURI(JSON.stringify(dehydrateFilters(filters)));
       updateRoute({ ...update, report: selectedReport?.id });
     }
   }, [
@@ -62,7 +57,6 @@ export function ReportOptions(props) {
     updateRoute,
     reportOptions.isReady,
     selectedReport,
-    isComparatorsEnabled,
     reportOptions,
   ]);
 
@@ -174,11 +168,7 @@ export function ReportOptions(props) {
                 <MetricsDrawer selectedMetrics={processedMetrics} handleSubmit={handleSubmit} />
               </Tabs.Item>
               <Tabs.Item>
-                {isComparatorsEnabled ? (
-                  <FiltersForm handleSubmit={handleSubmit} reportOptions={reportOptions} />
-                ) : (
-                  <AddFiltersSection handleSubmit={handleSubmit} reportOptions={reportOptions} />
-                )}
+                <FiltersForm handleSubmit={handleSubmit} reportOptions={reportOptions} />
               </Tabs.Item>
               {isCompareByEnabled && (
                 <Tabs.Item>
@@ -204,7 +194,7 @@ export function ReportOptions(props) {
             </Heading>
 
             {isComparatorsEnabled ? (
-              <ActiveFiltersV2
+              <ActiveFilters
                 filters={reportOptions.filters}
                 handleFilterRemove={handleFilterRemoveV2}
               />
@@ -223,7 +213,6 @@ export function ReportOptions(props) {
 
 const mapStateToProps = state => ({
   featureFlaggedMetrics: selectFeatureFlaggedMetrics(state),
-  isComparatorsEnabled: selectCondition(isAccountUiOptionSet('allow_report_filters_v2'))(state),
   isCompareByEnabled: selectCondition(isAccountUiOptionSet('allow_compare_by'))(state), //Comparing different filters
 });
